Extract diff-splice helpers into yaml-section-splice.ts
Move the new span model, structural value equality, and body
assembler out of the already-oversized yaml-section-values.ts into a
focused module, keeping that file on the parser/update API surface.
Behavior-free; the parse helpers stay put (they share local line
utilities).
